pyproject.toml
[tool.poetry]
name = "frenetix-rl"
version = "2024.1"
description = "Frenetix Reinforcement Learning Boosted Motion Planner"
readme = "README.md"
license = "LGPL-3.0-or-later"
authors = [
    "Rainer Trauth <rainer.trauth@tum.de>",
    "Alexander Hobmeier <alexander.hobmeier@tum.de>",
]
maintainers = [
    "Rainer Trauth <rainer.trauth@tum.de>",
]
packages = [
    { include = "frenetix_rl" },
]

[tool.poetry.dependencies]

python = ">=3.9,<3.11"

# internal, will be resolved when we publish it.
frenetix-motion-planner = "^2024.1"
commonroad-route-planner = "^2022.3"
commonroad-drivability-checker = "^2023.1"

frenetix = "^0.1.3"

numpy = "^1.24.2"
omegaconf = "^2.1.1"
scipy = "^1.9.3"
onnxruntime = "^1.16.3"
shapely = "^2.0.2"
wheel = "^0.38.4"
pandas = "^2.0.3"
gymnasium = "^0.29.1"

stable-baselines3 = { version = "^2.2.1", extras = ["extra"] }
sb3_contrib = "^2.2.1"
tensorboard = "^2.15.1"

commonroad-io = "==2023.2"
commonroad-vehicle-models = "^3.0.2"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=6.2.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
